requirements.txt
Flask
subprocess.run
google-generativeai